Add popup to Marker, add themes dir, cleanups
