+++ demo-app/src/POCMap.jsx
@@ -18,51 +18,52 @@ import {
 import { generateLocations, userLocation } from './utils/locations';
 import { ClusterPopup } from './ClusterPopup';
 import { getClusterStatus, getClusterSize } from './utils/status';
+import { hpeLeaflet } from './themes';
 
-const markerKinds = {
-  default: {
-    container: {
-      border: {
-        color: 'border-strong',
-        size: 'small',
-      },
-    },
-  },
-  critical: {
-    container: {
-      // any box props
-      background: 'background-front',
-      border: { color: 'status-critical', size: '3px' },
-    },
-    icon: <StatusCriticalSmall color="status-critical" size="13px" />,
-  },
-  warning: {
-    container: {
-      background: 'background-front',
-      border: { color: 'status-warning' },
-    },
-    icon: <StatusWarningSmall color="status-warning" size="13px" />,
-  },
-  good: {
-    container: {
-      background: 'background-front',
-      border: { color: 'status-ok' },
-    },
-    icon: <StatusGoodSmall color="status-ok" size="13px" />,
-  },
-  unknown: {
-    container: {
-      background: 'background-front',
-      border: { color: 'status-unknown' },
-    },
-    icon: <StatusUnknownSmall color="status-unknown" size="13px" />,
-  },
-};
+// const markerKinds = {
+//   default: {
+//     container: {
+//       border: {
+//         color: 'border-strong',
+//         size: 'small',
+//       },
+//     },
+//   },
+//   critical: {
+//     container: {
+//       // any box props
+//       background: 'background-front',
+//       border: { color: 'status-critical', size: '3px' },
+//     },
+//     icon: <StatusCriticalSmall color="status-critical" size="13px" />,
+//   },
+//   warning: {
+//     container: {
+//       background: 'background-front',
+//       border: { color: 'status-warning' },
+//     },
+//     icon: <StatusWarningSmall color="status-warning" size="13px" />,
+//   },
+//   good: {
+//     container: {
+//       background: 'background-front',
+//       border: { color: 'status-ok' },
+//     },
+//     icon: <StatusGoodSmall color="status-ok" size="13px" />,
+//   },
+//   unknown: {
+//     container: {
+//       background: 'background-front',
+//       border: { color: 'status-unknown' },
+//     },
+//     icon: <StatusUnknownSmall color="status-unknown" size="13px" />,
+//   },
+// };
 
-const hpeLeaflet = {
-  pin: { ...markerKinds },
-  cluster: { ...markerKinds },
-};
+// const hpeLeaflet = {
+//   pin: { ...markerKinds },
+//   cluster: { ...markerKinds },
+// };
 
 function POCMap() {
   const [geolocation, setGeolocation] = useState();
