+++ demo-app/src/themes/index.js
@@ -1,0 +1,2 @@
+export * from './hpeLeaflet';
+export * from './generic';
